--- tests/azure/terraform/aks/terraform.tfvars
@@ -1,1 +1,0 @@
-resource_group = "test-fluxes"
