+++ tests/integration/kustomization.yaml
@@ -15,4 +15,7 @@ patches:
     - op: add
       path: /spec/template/spec/containers/0/args/-
       value: --azure-autologin-for-acr
+    - op: add
+      path: /spec/template/spec/containers/0/args/-
+      value: --aws-autologin-for-ecr
