+++ tests/integration/terraform/aks/acr.tf
