+++ tests/integration/terraform/aks/main.tf
@@ -8,10 +8,6 @@ locals {
     name = "flux-test-${random_pet.suffix.id}"
 }
 
-variable "resource_group" {
-    type = string
-}
-
 variable "region" {
     type = string
     default = "eastus"
