+++ tests/integration/terraform/aks/output.tf
@@ -2,7 +2,7 @@ output "aks_kubeconfig" {
     value = azurerm_kubernetes_cluster.default.kube_config_raw
 }
 
-output "acr_repository_url" {
+output "repository_url" {
     value = azurerm_container_registry.acr.login_server
 }
