+++ tests/integration/terraform/eks/ecr.tf
@@ -1,0 +1,4 @@
+resource "aws_ecr_repository" "testrepo" {
+  name                 = "flux-test-repo-${random_pet.suffix.id}"
+  image_tag_mutability = "MUTABLE"
+}
