+++ tests/integration/terraform/eks/eks.tf
@@ -1,0 +1,45 @@
+module "eks" {
+  source  = "terraform-aws-modules/eks/aws"
+  version = "~> 18"
+
+  cluster_name    = local.name
+  cluster_version = "1.22"
+
+  # Maybe don't need any of these?
+  cluster_endpoint_private_access = true
+  cluster_endpoint_public_access  = true
+
+  cluster_addons = {
+    coredns = {
+      resolve_conflicts = "OVERWRITE"
+    }
+    kube-proxy = {}
+    vpc-cni = {
+      resolve_conflicts = "OVERWRITE"
+    }
+  }
+
+  vpc_id     = module.vpc.vpc_id
+  subnet_ids = module.vpc.private_subnets
+
+  eks_managed_node_group_defaults = {
+    disk_size      = 50
+    instance_types = ["t2.medium"]
+  }
+
+  eks_managed_node_groups = {
+    blue = {}
+    green = {
+      min_size     = 1
+      max_size     = 2
+      desired_size = 1
+
+      instance_types = ["t2.medium"]
+      capacity_type  = "SPOT"
+    }
+  }
+}
+
+data "aws_caller_identity" "current" {}
+
+data "aws_region" "current" {}
