+++ tests/integration/terraform/eks/main.tf
@@ -1,0 +1,7 @@
+provider "aws" {}
+
+resource "random_pet" "suffix" {}
+
+locals {
+    name = "flux-test-${random_pet.suffix.id}"
+}
