+++ tests/tftestenv/azure.go
@@ -1,0 +1,42 @@
+/*
+Copyright 2022 The Flux authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package tftestenv
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+
+	tfjson "github.com/hashicorp/terraform-json"
+)
+
+// CreateAKSKubeConfig constructs kubeconfig for an AKS cluster from the terraform state output at the
+// given kubeconfig path.
+func CreateAKSKubeConfig(ctx context.Context, state map[string]*tfjson.StateOutput, kcPath string) error {
+	kubeconfigYaml, ok := state["aks_kubeconfig"].Value.(string)
+	log.Println(kubeconfigYaml)
+	if !ok || kubeconfigYaml == "" {
+		return fmt.Errorf("failed to obtain kubeconfig from tf output")
+	}
+
+	f, err := os.Create(kcPath)
+	if err != nil {
+		return err
+	}
+	_, err = fmt.Fprint(f, kubeconfigYaml)
+	return err
+}
